pyproject.toml
[tool.poetry]
name = "cheatsheet-maker"
version = "0.1.0"
description = ""
authors = ["Waldoh <waldoh@gptcoder.pro>"]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.27.2"
openai = "^0.28.1"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
